requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.1.6
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
psycopg==3.0.10
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
